apps/agent/requirements.txt
livekit-agents[deepgram,silero,turn-detector]~=1.5
python-dotenv~=1.1
